+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/ServerControllerCodeGeneratorTest.kt
@@ -207,6 +207,18 @@ internal class ServerControllerCodeGeneratorTest {
         assertEquals(
             """
                 @org.springframework.web.bind.`annotation`.RestController
+                @org.springframework.web.bind.`annotation`.CrossOrigin(
+                  origins = ["*"],
+                  allowedHeaders = ["Content-Type", "Authorization", "Content-Length", "X-Requested-With"],
+                  methods = [
+                      org.springframework.web.bind.annotation.RequestMethod.GET,
+                      org.springframework.web.bind.annotation.RequestMethod.DELETE,
+                      org.springframework.web.bind.annotation.RequestMethod.OPTIONS,
+                      org.springframework.web.bind.annotation.RequestMethod.HEAD,
+                      org.springframework.web.bind.annotation.RequestMethod.POST,
+                      org.springframework.web.bind.annotation.RequestMethod.PUT
+                  ]
+                )
                 public class CreateController(
                   private val `delegate`: com.wutsi.test.`delegate`.CreateDelegate
                 ) {